requirements.txt
numpy==1.23.3
opencv-python==4.7.0.72
pandas==1.5.0
pandas-stubs==1.5.1.221024
Pillow==9.2.0
streamlit==1.13.0
torch==2.0.0
torchvision==0.15.1
transformers==4.24.0
